Problem with bottom edge of generated level?
I put in two TODO: QUESTION comments for Vanessa @TheHedgeify regarding
the Mario level generation. I copied some of this code to a separate
Mario branch in order to generate some levels, but when I ran it there
was an off-by-one error because the bottom row was being ignored. I
fixed it in my code by changing it as suggested, but didn't want to
"fix" anything that wasn't broken, hence tagging you in this commit.