Changed path of json file being pointed to
The MainRun class was looking for the json file in a subdir on Vanessa's
computer, but I changed it to look to the project root since that is
where we are currently storing that file.

I also changed GANEvaluator in the same way, though this doesn't really
matter since that class doesn't do much currently.