Added cmatest.marioobjectives package with objective function interface and example with A* agent
Once the levelFromLatentVector method in the MarioLevelObjective class
is completed, we should have an easy time creating various objective
functions for the levels the GAN produces. This currently incomplete
method is meant to be the one place where we ask the GAN to give us a
level.

Also added example AStarAgentProgressMarioObjective class that extends
the AgentProgressMarioObjective, which rates a level based on how well a
particular agent performs in it. The was this is written, it should be
very easy to evaluate other agents in the GAN levels as well.