Added support for the Bullet Bill cannon launchers
These were not described in the VGLC json description summary, but they
were present in the data. However, there is a bit of a trick regarding
their representation, since what is a single character in VGLC is
actually two different sprites in Infinite Mario. This may require some
fixing, but I think it is purely an aesthetic difference.