Adds LARGE training set
This new training set uses all the Mario levels we have, but also the
Mario 2 (Japan) levels and the Mario Land (Gameboy) levels.
Additionally, these level representations include the B and b
characters, which are related to Bullet Bill cannons ... though we might
want to remvoe this extra tile type to simplify learning.