Closed the loop on the Java side!
I've closed the loop from the Java side, meaning that the Java version of CMA-ES can be launched, but it will use a perpetually running Python process (containing the model) to generate Mario levels from latent vectors. The level Strings get returned to Java, which then evaluates the level using the AStar Mario agent.

To see how this works, execute CMAMarioSolver

Note that there is still a bit of clean up to be done. I think the code is pretty messy. Obviously, it is using the bad model that doesn't really work. This is quite obvious because every Mario evaluation is visually displayed, and the levels aren't great. It is useful to actually see the levels being generated/evaluated, but for speed we should do the evaluations without visual display. Also, currently the code sends one latent vector to the Python program at a time, but it would be faster to send them in batches. Lastly, the evlauation based on the AStar agent is pretty specific ... we need to generalize the evaluation options.

Sorry to @ljialin for stepping in to take this over ... I was so excited that I could actually get the Python Keras code to run on my machine despite using Windows, that I just keep working on it.

However, I'm still going to take a shot at launching the evaluation through Python using the Python version of CMA-ES. I think that evaluating the population in parellel might be easier this way, and over all it might be faster. However, at least we have something we can run now if the GAN problems are solved.

Hopefully, I can also find some time to clean up this code. There is currently some dead code here and there. Some of the dead code seems to be because @ljialin and I were doing the same things in different ways, and some of it seems to be left-over code from the GVG-AI process communication code that was copied from elsewhere.