Removed simulateOneLevel call
Got home and tested my recent refactoring, and noticed that at some point an unnecessary call to simulateOneLevel was added. I removed this ... we don't want  to simulate a level before the GAN sends one.

However, I did notice that Robin Baumgarten's A* agent sometimess has trouble with or even crashes on the weird levels the GAN is producing ... we need a way to catch these exceptions and continue on regardless.