Delete incomplete file for agent-based evolution from Python
At one point we were working on doing the A* agent based evolution in both Python and Java. This was in part because at one point neither one worked, and because we were concerned about speed. However, I think we can get rid of this file now. Java works fine for the agent-based experiments, so we don't need this incomplete file around.