Delete Adam's original Keras code?
This code is all associated with Adam's initial GAN model that we ended up not using in the end. The ipynb notebook files were used to create the model, and generator.py was used to run the model by our Java code. Also, generator.h5 and generator.json were the saved weights and achitecture of his model. I don't think we should keep these files around in the codebase anymore, both because they don't work, and because we didn't use them in the actual paper experiments. It is best to get rid of them, or at least move them into some kind of subdirectory for old/unused code.