Fixed inconsistent whitespace
When I tried running this on a Mac in complained about inconsistent mixing of spaces and tabs in the whitespace, so I fixed it.

@thehedgeify Please confirm that this still runs properly on your machine.

I also modified evaluateLevel.java so that it automatically loads one of your newly added GANs by default if no command line parameters are specified. The file also automatically lets a human play the level afterward, but enabling this required me to uncomment (and change) some code ... perhaps you don't want this to execute by default for your testing purposes, but it would probably be nice if people using the code for the first time can quickly and easily play levels generated by the GAN.