Changes to the Python files to work with the latest PyTorch
PyTorch now disallows the use of periods "." in layer/module names.
Therefore, all names in the DCGAN model had each period "." replaced
with a colon ":". However, this creates a backwards compatibility issue
with the old models we trained, so in generator_ws I have to account for
this: first I load the saved model (as an OrderedDict) and then loop
through to create a new OrderedDict that has the parameters loaded from
the saved model, but the module/layer labels that use colons instead of
periods