Make sure that old GAN is launched by evolution code by default
If a user launches CMAMarioSolver without any command line parameters,
it will now correctly load the old GECCO 2018 Mario GAN and use it to
evolve levels. However, I suppose we are using the new LevelParser
rather than the old LevelParser, so I suppose the levels may not be
evolved in exactly the same way ... but the code seems to run.