Load models in new version of PyTorch
The new versions of PyTorch do not like periods in the names. I already made code that loads old models with periods and converts them to underscores. However, if you train a new model, it will have the underscores. This change makes it possible to load those models.

I have verified that this change works on the Python end via the command line, but I have not successfully used a model trained in this way in conjunction with the Java evolution code.

This is related to issue #2